requirements.txt
# Core Database
sqlmodel>=0.0.8
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Step 1: Facebook Scraping
requests>=2.28.0
tldextract>=3.4.0

# Step 2: Traffic Analysis
aiohttp>=3.8.0

# Environment Variables
python-dotenv>=1.0.0
